pyproject.toml
[project]
name = "intelligent-darts"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "python-multipart",
    "ultralytics",
    "opencv-python-headless",
    "boto3",
]

[dependency-groups]
dev = [
    "ty>=0.0.12", "apx==0.2.6",
]

[tool.apx.metadata]
app-name = "intelligent-darts"
app-slug = "intelligent_darts"
app-entrypoint = "intelligent_darts.backend.app:app"
api-prefix = "/api"
metadata-path = "src/intelligent_darts/_metadata.py"

[tool.apx.ui]
root = "src/intelligent_darts/ui"

[tool.apx.ui.registries]
"@animate-ui" = "https://animate-ui.com/r/{name}.json"
"@ai-elements" = "https://registry.ai-sdk.dev/{name}.json"
"@svgl" = "https://svgl.app/r/{name}.json"


[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
fallback-version = "0.0.0"

[tool.uv]

[[tool.uv.index]]
name = "apx-index"
url = "https://databricks-solutions.github.io/apx/simple"

[tool.uv.sources]

[tool.uv.sources.apx]
index = "apx-index"

[tool.hatch.build.hooks.version]
path = "src/intelligent_darts/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/intelligent_darts/__dist__", "src/intelligent_darts/_metadata.py"]
exclude = ["src/intelligent_darts/ui"]
